Remove circular dependency
VTMS-2743
